+++ src/main/java/com/esp/service/GenericService.java
@@ -8,7 +8,7 @@ public interface GenericService<T> {
 	
 public void add(T t);
 
-public T fetch(BigDecimal id); 
+public T fetch(int id); 
 
 public List<T> fetchByParam(Object obj);
